+++ .idea/libraries/support_annotations_22_2_1.xml
@@ -1,12 +1,11 @@
 <component name="libraryTable">
   <library name="support-annotations-22.2.1">
     <CLASSES>
-      <root url="jar://D:/software/Android/sdk/extras/android/m2repository/com/android/support/support-annotations/22.2.1/support-annotations-22.2.1.jar!/" />
+      <root url="jar://$USER_HOME$/software/android-sdk-linux/extras/android/m2repository/com/android/support/support-annotations/22.2.1/support-annotations-22.2.1.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES>
-      <root url="jar://D:/software/Android/sdk/extras/android/m2repository/com/android/support/support-annotations/22.2.1/support-annotations-22.2.1-sources.jar!/" />
-      <root url="jar://$PROJECT_DIR$/../../Androidstudio/sdk/android-sdk-linux/extras/android/m2repository/com/android/support/support-annotations/22.2.1/support-annotations-22.2.1-sources.jar!/" />
+      <root url="jar://$USER_HOME$/software/android-sdk-linux/extras/android/m2repository/com/android/support/support-annotations/22.2.1/support-annotations-22.2.1-sources.jar!/" />
     </SOURCES>
   </library>
 </component>
