+++ app/src/main/AndroidManifest.xml
@@ -16,6 +16,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".TowDegree"
+            android:label="@string/title_activity_tow_degree" >
+        </activity>
     </application>
 
 </manifest>
