+++ app/src/main/java/com/exercise/swd3/two_degree/MainActivity.java
@@ -1,5 +1,6 @@
 package com.exercise.swd3.two_degree;
 
+import android.content.Intent;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.view.Menu;
@@ -10,59 +11,17 @@ import android.widget.TextView;
 
 
 public class MainActivity extends ActionBarActivity {
-    private Button meBtn = null;
-    private Button twoDegreeBtn = null;
-    private Button dynamicBtn = null;
-    private Button informationBtn = null;
-    private TextView textView = null;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        meBtn = (Button)findViewById(R.id.meBtn);
-        twoDegreeBtn = (Button)findViewById(R.id.two_degreeBtn);
-        dynamicBtn = (Button)findViewById(R.id.dynamicBtn);
-        informationBtn = (Button)findViewById(R.id.informationBtn);
-        textView = (TextView)findViewById(R.id.testView);
-        meBtn.setOnClickListener(new View.OnClickListener() {
+        setContentView(R.layout.log_in);
+
+        findViewById(R.id.loginBtn).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                textView.setText(R.string.me_text);
-                meBtn.setTextColor(getResources().getColor(R.color.blue));
-                twoDegreeBtn.setTextColor(getResources().getColor(R.color.black));
-                dynamicBtn.setTextColor(getResources().getColor(R.color.black));
-                informationBtn.setTextColor(getResources().getColor(R.color.black));
-            }
-        });
-        twoDegreeBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                textView.setText(R.string.two_degree_text);
-                meBtn.setTextColor(getResources().getColor(R.color.black));
-                twoDegreeBtn.setTextColor(getResources().getColor(R.color.blue));
-                dynamicBtn.setTextColor(getResources().getColor(R.color.black));
-                informationBtn.setTextColor(getResources().getColor(R.color.black));
-            }
-        });
-        dynamicBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                textView.setText(R.string.dynamic_text);
-                dynamicBtn.setTextColor(getResources().getColor(R.color.blue));
-                meBtn.setTextColor(getResources().getColor(R.color.black));
-                twoDegreeBtn.setTextColor(getResources().getColor(R.color.black));
-                informationBtn.setTextColor(getResources().getColor(R.color.black));
-            }
-        });
-        informationBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                textView.setText(R.string.information_text);
-                meBtn.setTextColor(getResources().getColor(R.color.black));
-                twoDegreeBtn.setTextColor(getResources().getColor(R.color.black));
-                dynamicBtn.setTextColor(getResources().getColor(R.color.black));
-                informationBtn.setTextColor(getResources().getColor(R.color.blue));
+                startActivity(new Intent(MainActivity.this,TowDegree.class));
             }
         });
