+++ app/src/main/res/layout/activity_tow_degree.xml
@@ -1,0 +1,71 @@
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingBottom="0dp" tools:context=".MainActivity">
+    <LinearLayout
+        android:orientation="vertical"
+        android:layout_width ="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        >
+        <LinearLayout
+            android:orientation="vertical"
+            android:layout_width ="match_parent"
+            android:layout_height="match_parent"
+            android:layout_alignParentBottom="true"
+            android:layout_alignParentLeft="true"
+            android:layout_alignParentStart="true"
+            android:layout_weight="1"
+            android:weightSum="1">
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/information_text"
+                android:id="@+id/testView"
+                android:textSize="30dp"
+                android:layout_marginTop="200dp"
+                android:layout_gravity="center_horizontal" />
+        </LinearLayout>
+        <LinearLayout
+            android:orientation="horizontal"
+            android:layout_width ="match_parent"
+            android:layout_height="match_parent"
+            android:layout_alignParentBottom="true"
+            android:layout_alignParentLeft="true"
+            android:layout_alignParentStart="true"
+            android:layout_weight="9"
+            >
+            <Button android:layout_width="wrap_content"
+                android:layout_height="fill_parent"
+                android:text="@string/information_button"
+                android:id="@+id/informationBtn"
+                android:layout_weight="1"
+                android:textColor="#c9124afc"
+                android:background="#FFFFFF"/>
+            <Button android:layout_width="wrap_content"
+                android:layout_height="fill_parent"
+                android:text="@string/two_degree_button"
+                android:id="@+id/two_degreeBtn"
+                android:layout_weight="1"
+                android:textColor="#000000"
+                android:background="#FFFFFF"/>
+            <Button android:layout_width="wrap_content"
+                android:layout_height="fill_parent"
+                android:text="@string/dynamic_button"
+                android:id="@+id/dynamicBtn"
+                android:textColor="#000000"
+                android:background="#FFFFFF"
+                android:layout_weight="1"/>
+            <Button android:id="@+id/meBtn"
+                android:layout_width="wrap_content"
+                android:layout_height="fill_parent"
+                android:layout_weight="1"
+                android:textColor="#255"
+                android:background="#FFFFFF"
+                android:text="@string/me_button"/>
+        </LinearLayout>
+    </LinearLayout>
+
+</RelativeLayout>
