+++ local.properties
@@ -7,5 +7,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sat Jul 25 09:11:17 CST 2015
-sdk.dir=/home/swd3/Android/Androidstudio/sdk/android-sdk-linux
+#Sun Aug 09 16:58:10 CST 2015
+sdk.dir=/home/swd3/software/android-sdk-linux
